Added Play button functionality
