+++ package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-router-dom": "^6.21.1",
     "react-scripts": "5.0.1"
   },
   "devDependencies": {
